Update task_definition_parameters.md
The existing example for health check is:
[ "CMD-SHELL", "curl -f http://localhost/ || exit 1" ]
But if a customer/user copies and pastes this into the health check section in the Console, this will lead to health checks being always unknown. The above example is what the task definition looks like once a health check has been implemented. I've edited the documentation to reflect this change, so that customers understand what the input should be.