Update service-discovery.md
This required clarification, led to assumptions while reading. In my testing the A records of the tasks are returned, not the A record of the load balancer, for eg:

;; QUESTION SECTION:
;web.prod.local.			IN	A

;; ANSWER SECTION:
web.prod.local.		10	IN	A	172.31.9.227
web.prod.local.		10	IN	A	172.31.6.180

The above are the private IPs of two tasks in a service with a private namespace.

Can you confirm these suggested changes are correct?