Change CPU explanation
The docker docs for CPU shares state:

"It prioritizes container CPU resources for the available CPU cycles. It does not guarantee or reserve any specific CPU access."
https://docs.docker.com/config/containers/resource_constraints/#configure-the-default-cfs-scheduler

Which is in conflict with ECS docs which currently state:
"The number of cpu units to reserve for the container. This parameter maps to CpuShares in the Create a container section of the Docker Remote API and the --cpu-shares option to docker run."

I propose we change this to state that this is an ECS Agent abstraction over the concept of cpu-shares to avoid confusion.