Remove reference to "CodeDeploy"
The line I am editing is 
>> The `imagedefinitions.json` file is required for the CodeDeploy ECS job worker\.

To:

 The `imagedefinitions.json` file is required for the ECS job worker\.
CodeDeploy ECS job worker is for Blue/Green deployment not the standard one. The CodeDeploy ECS job worker needs file "imageDetail.json" for Amazon ECS Blue/Green Deployment Actions and not "imagedefinitions.json". When you mention the word CodeDeploy in this tutorial, it causes confusion that this is a Blue/Green deployment which this is not.